renamed files to match project. Added getBrands funct
